move featured songs to own component
