+++ src/components/ArtistPortfolio/styles.css
@@ -1,41 +1,4 @@
 
-ol {
-    margin-left: 2em;
-}
-
-li {
-    display: flex;
-    align-items: center;
-    /* margin: 20px;    */
-    margin-bottom: 1em;
-    padding: 1em;
-    background-color:  white;
-    color: black;
-    border-radius: 2px;
-}
-.song-title {
-    justify-self: flex-start;
-    flex: 1;
-}
-
-.song-plays, .fav-btn, .lyrics-btn{
-    margin-left: 1em;
-    justify-self: flex-end;
-}
-
-.btn {
-    transition: all 0.2s;
-    cursor: pointer;
-    border-radius: 100%;
-    padding: 5px;
-}
-
-.btn:hover {
-    transform: scale(1.4);
-}
-
-
-
 .album-container {
     /* max-width: 100%; */
     /* display: flex; */
